Moving around some code [skip hazardlib][demos]
